# services/user/docs/swagger.yaml
basePath: /api/v1/user
definitions:
  errorx.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 音频 AI 平台用户服务接口文档，包含用户登录、注册、个人信息、家庭管理、设备分享、下载管理、实名认证等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 用户服务 API 文档
  version: "1.0"
paths:
...
  /user/content/download/stream:
    get:
      consumes:
      - application/json
      description: 流式传输私有格式内容，同时保存到本地
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 流式下载
      tags:
      - 下载服务
...
  /user/device/share/create:
    post:
      consumes:
      - application/json
      description: 创建设备共享，将设备权限分享给其他用户
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 创建设备共享
      tags:
      - 设备分享
  /user/device/share/detail:
    get:
      consumes:
      - application/json
      description: 查询指定设备共享的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 共享详情
      tags:
      - 设备分享
...
  /user/device/share/sent:
    get:
      consumes:
      - application/json
      description: 查询当前用户发出的所有设备共享记录
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 发出的共享列表
      tags:
      - 设备分享
...
  /user/download/records:
    get:
      consumes:
      - application/json
      description: 查询当前用户的下载记录列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      security:
      - BearerAuth: []
      summary: 下载记录列表
      tags:
      - 下载服务
...
  /user/login/sessions/kick:
    post:
      consumes:
      - application/json
      description: 强制吊销指定登录会话，使该会话立即失效
      parameters:
      - description: 请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      summary: 吊销会话
      tags:
      - 登录会话
...
  /user/oauth/google/callback:
    get:
      consumes:
      - application/json
      description: Google OAuth 授权回调，自动创建或绑定用户
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      summary: Google OAuth 回调
      tags:
      - OAuth
...
  /user/oauth/wechat/callback:
    get:
      consumes:
      - application/json
      description: 微信 OAuth 授权回调，自动创建或绑定用户
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      summary: 微信 OAuth 回调
      tags:
      - OAuth
...
  /user/password/reset:
    post:
      consumes:
      - application/json
      description: 忘记密码时，通过邮箱/手机验证码重置密码
      parameters:
      - description: 重置密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      summary: 重置密码
      tags:
      - 密码管理
...
  /user/register:
    post:
      consumes:
      - application/json
      description: 新用户注册，支持邮箱或手机号注册，需先发验证码
      parameters:
      - description: 注册请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/errorx.Response'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/errorx.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorx.Response'
        "409":
          description: 账号已存在
          schema:
            $ref: '#/definitions/errorx.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/errorx.Response'
      summary: 用户注册
      tags:
      - 用户认证
...
securityDefinitions:
  BearerAuth:
    description: 请输入 Bearer Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"