# services/user/internal/handler/list_download_records_handler.go
package handler

import (
	"net/http"

	"github.com/jacklau/audio-ai-platform/services/user/internal/logic"
	"github.com/jacklau/audio-ai-platform/services/user/internal/pkg/util/ctxuser"
	"github.com/jacklau/audio-ai-platform/services/user/internal/svc"
	"github.com/jacklau/audio-ai-platform/services/user/internal/types"
	"github.com/zeromicro/go-zero/rest/httpx"
)

// listDownloadRecordsHandler 查询用户下载记录列表处理器
// @Summary      下载记录列表
// @Description  查询当前用户的下载记录列表
// @Tags         下载服务
// @Accept       json
// @Produce      json
// @Param        page  query  int  false  "页码"
// @Param        page_size  query  int  false  "每页数量"
// @Param        status  query  string  false  "状态筛选"
// @Success      200  {object}  types.ListDownloadRecordsResp  "成功"
// @Failure      400  {object}  errorx.Response  "参数错误"
// @Failure      401  {object}  errorx.Response  "未登录"
// @Failure      500  {object}  errorx.Response  "服务器错误"
// @Router       /user/download/records [get]
// @Security     BearerAuth
// GET /api/v1/user/download/records
// 用途：查询当前用户的下载记录列表，支持分页、状态筛选、关键词搜索
func listDownloadRecordsHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if r.Method != http.MethodGet {
			httpx.WriteJson(w, http.StatusMethodNotAllowed, map[string]interface{}{
				"code": 405,
				"msg":  "仅支持 GET",
				"data": nil,
			})
			return
		}

		userID := ctxuser.ParseUserID(r.Context())
		if userID == 0 {
			httpx.WriteJson(w, http.StatusUnauthorized, map[string]interface{}{
				"code": 401,
				"msg":  "未登录",
				"data": nil,
			})
			return
		}

		var req types.ListDownloadRecordsReq
		if err := httpx.Parse(r, &req); err != nil {
			httpx.WriteJson(w, http.StatusBadRequest, map[string]interface{}{
				"code": 400,
				"msg":  "参数错误",
				"data": nil,
			})
			return
		}

		l := logic.NewListDownloadRecordsLogic(r.Context(), svcCtx)
		resp, err := l.ListDownloadRecords(userID, &req)
		if err != nil {
			httpx.WriteJson(w, http.StatusInternalServerError, map[string]interface{}{
				"code": 500,
				"msg":  err.Error(),
				"data": nil,
			})
			return
		}

		httpx.WriteJson(w, http.StatusOK, map[string]interface{}{
			"code": 0,
			"msg":  "成功",
			"data": resp,
		})
	}
}